reconciling the naming for postgresql products because there is cosmos db as well as azure db postgresql flex.
