+++ package/.env.sample.tests
@@ -2,7 +2,11 @@
 COSMOS_DB_MONGO_URI=""
 DB_NAME=""
 COLLECTION_NAME=""
-# CosmosDB postgres 
-COSMOSDB_POSTGRES_CONN_STRING=""
+
+# PostgreSQL
+COSMOS_DB_POSTGRESQL_CONN_STRING=""
 TABLE_NAME=""
 
+# PostgreSQL flex
+POSTGRESQL_FLEX_CONN_STRING=""
+TABLE_NAME=""
