+++ package/pfazuredb/tools/postgresqlpgvector.py
@@ -20,7 +20,7 @@ def vectorsearch(
     import json
 
     # establish connection
-    pgconnection = psycopg2.connect(connection.secrets["AZURE_COSMOSDB_POSTGRES_CONN_STRING"])
+    pgconnection = psycopg2.connect(connection.secrets["POSTGRESQL_CONN_STRING"])
     register_vector(pgconnection)
 
     if vectorsearch_method == "L2":
