+++ package/setup.py
@@ -5,7 +5,7 @@ PACKAGE_NAME = "pf-azuredb"
 setup(
     name=PACKAGE_NAME,
     author="Hossein K. Heris and applied ai team@azure data",
-    version="1.0.0",
+    version="1.0.1",
     description="Package for use in promptflow for vector search in azure db",
     packages=find_packages(),
     entry_points={
